refactor: update dialog and stored item types for improved consistency
- Replaced `StoredItem` with `StoredItemMapped` in `dialogs.ts`, `MessageFile.vue`, and `CyberlinkResult.vue` to ensure uniformity across components.
- Adjusted the `MessageItem.vue` to handle `stored_items` as `StoredItemMapped` and updated related logic accordingly.
- Introduced a new composable `useDialogView` to encapsulate dialog-related functionalities, enhancing code organization.
- Refactored `useLlmDialog` to integrate with the new dialog view logic, streamlining the assistant's interaction with dialogs.
- Cleaned up unused code and improved type handling in various components for better maintainability.
