+++ src/components/MessageFile.vue
@@ -12,13 +12,13 @@
 
 <script setup lang="ts">
 import { useQuasar } from 'quasar'
-import { StoredItem } from '@/services/supabase/types'
+import { StoredItemMapped } from '@/services/supabase/types'
 import ViewFileDialog from './ViewFileDialog.vue'
 import { computed } from 'vue'
 import { codeExtensions } from 'src/utils/values'
 
 const props = defineProps<{
-  file: StoredItem,
+  file: StoredItemMapped,
   removable?: boolean
 }>()
